requirements.txt
# Existing requirements
torch
torchvision
timm
transformers
datasets
accelerate
albumentations
opencv-python
pandas
scikit-learn
matplotlib
seaborn
tqdm
onnx
onnxruntime
torchmetrics
catalyst
Flask>=2.3.0
Flask-CORS
Werkzeug
Pillow
numpy
python-dotenv
fastapi
uvicorn
python-multipart
pydantic
httpx
faiss-cpu
python-dotenv
